test(config): fix misleading test mocks that silently masked failures
