+++ lib/workers/global/config/parse/index.spec.ts
@@ -4,6 +4,7 @@ import { CONFIG_PRESETS_INVALID } from '../../../../constants/error-messages';
 import { getCustomEnv } from '../../../../util/env';
 import { getParentDir, readSystemFile } from '../../../../util/fs';
 import getArgv from './__fixtures__/argv';
+import * as _fileConfigParser from './file';
 import * as _hostRulesFromEnv from './host-rules-from-env';
 import * as httpMock from '~test/http-mock';
 
@@ -11,8 +12,10 @@ vi.mock('../../../../modules/datasource/npm');
 vi.mock('../../../../util/fs');
 vi.mock('../../../../config/decrypt');
 vi.mock('./host-rules-from-env');
+vi.mock('./file');
 
 const decrypt = vi.mocked(_decrypt);
+const fileConfigParser = vi.mocked(_fileConfigParser);
 
 const { hostRulesFromEnv } = vi.mocked(_hostRulesFromEnv);
 
@@ -34,12 +37,14 @@ describe('workers/global/config/parse/index', () => {
     });
 
     it('supports token in env', async () => {
+      fileConfigParser.getConfig.mockResolvedValue({});
       const env: NodeJS.ProcessEnv = { ...defaultEnv, RENOVATE_TOKEN: 'abc' };
       const parsedConfig = await configParser.parseConfigs(env, defaultArgv);
       expect(parsedConfig).toContainEntries([['token', 'abc']]);
     });
 
     it('supports token in CLI options', async () => {
+      fileConfigParser.getConfig.mockResolvedValue({});
       defaultArgv = defaultArgv.concat([
         '--token=abc',
         '--pr-footer=custom',
@@ -57,6 +62,7 @@ describe('workers/global/config/parse/index', () => {
     });
 
     it('supports forceCli', async () => {
+      fileConfigParser.getConfig.mockResolvedValue({});
       defaultArgv = defaultArgv.concat(['--force-cli=false']);
       const env: NodeJS.ProcessEnv = {
         ...defaultEnv,
@@ -71,6 +77,7 @@ describe('workers/global/config/parse/index', () => {
     });
 
     it('sets customEnvVariables', async () => {
+      fileConfigParser.getConfig.mockResolvedValue({});
       const env: NodeJS.ProcessEnv = {
         ...defaultEnv,
         RENOVATE_TOKEN: 'abc',
@@ -84,24 +91,29 @@ describe('workers/global/config/parse/index', () => {
     });
 
     it('supports config.force', async () => {
-      const configPath = upath.join(__dirname, '__fixtures__/with-force.js');
-      const env: NodeJS.ProcessEnv = {
-        ...defaultEnv,
-        RENOVATE_CONFIG_FILE: configPath,
-      };
-      const parsedConfig = await configParser.parseConfigs(env, defaultArgv);
+      fileConfigParser.getConfig.mockResolvedValueOnce({
+        token: 'abcdefg',
+        force: {
+          schedule: ['* * * * 0.6'],
+        },
+      });
+      const parsedConfig = await configParser.parseConfigs(
+        defaultEnv,
+        defaultArgv,
+      );
       expect(parsedConfig).toContainEntries([
         ['token', 'abcdefg'],
         [
           'force',
           {
-            schedule: null,
+            schedule: ['* * * * 0.6'],
           },
         ],
       ]);
     });
 
     it('reads private key from file', async () => {
+      fileConfigParser.getConfig.mockResolvedValue({});
       const privateKeyPath = upath.join(__dirname, '__fixtures__/private.pem');
       const privateKeyPathOld = upath.join(
         __dirname,
@@ -120,6 +132,7 @@ describe('workers/global/config/parse/index', () => {
     });
 
     it('supports Bitbucket username/password', async () => {
+      fileConfigParser.getConfig.mockResolvedValue({});
       defaultArgv = defaultArgv.concat([
         '--platform=bitbucket',
         '--username=user',
@@ -137,6 +150,7 @@ describe('workers/global/config/parse/index', () => {
     });
 
     it('massages trailing slash into endpoint', async () => {
+      fileConfigParser.getConfig.mockResolvedValue({});
       defaultArgv = defaultArgv.concat([
         '--endpoint=https://github.renovatebot.com/api/v3',
       ]);
@@ -145,12 +159,14 @@ describe('workers/global/config/parse/index', () => {
     });
 
     it('parses global manager config', async () => {
+      fileConfigParser.getConfig.mockResolvedValue({});
       defaultArgv = defaultArgv.concat(['--detect-global-manager-config=true']);
       const parsed = await configParser.parseConfigs(defaultEnv, defaultArgv);
       expect(parsed.npmrc).toBeNull();
     });
 
     it('parses host rules from env', async () => {
+      fileConfigParser.getConfig.mockResolvedValue({});
       defaultArgv = defaultArgv.concat(['--detect-host-rules-from-env=true']);
       hostRulesFromEnv.mockReturnValueOnce([{ matchHost: 'example.org' }]);
       const parsed = await configParser.parseConfigs(defaultEnv, defaultArgv);
@@ -158,6 +174,7 @@ describe('workers/global/config/parse/index', () => {
     });
 
     it('env dryRun = true replaced to full', async () => {
+      fileConfigParser.getConfig.mockResolvedValue({});
       const env: NodeJS.ProcessEnv = {
         ...defaultEnv,
         RENOVATE_DRY_RUN: 'true',
@@ -167,32 +184,28 @@ describe('workers/global/config/parse/index', () => {
     });
 
     it('cli dryRun = true replaced to full', async () => {
+      fileConfigParser.getConfig.mockResolvedValue({});
       defaultArgv = defaultArgv.concat(['--dry-run=true']);
       const parsed = await configParser.parseConfigs(defaultEnv, defaultArgv);
       expect(parsed).toContainEntries([['dryRun', 'full']]);
     });
 
     it('resolves global presets', async () => {
+      fileConfigParser.getConfig.mockResolvedValue({
+        globalExtends: ['http://example.com/config.json', ':pinVersions'],
+        dryRun: 'extract',
+      });
       // The remote preset defined in globalExtends of the config file
       httpMock
         .scope('http://example.com/')
         .get('/config.json')
         .reply(200, { repositories: ['g/r1', 'g/r2'], druRun: 'full' });
 
-      // Mock the default config file to return a preset
-      const defaultConfig = upath.join(__dirname, '__fixtures__/default.js');
-      vi.doMock(defaultConfig, () => ({
-        default: {
-          globalExtends: ['http://example.com/config.json', ':pinVersions'],
-          dryRun: 'extract',
-        },
-      }));
-
       const parsedConfig = await configParser.parseConfigs(
         defaultEnv,
         defaultArgv,
       );
-      vi.doUnmock(defaultConfig);
+      // vi.doUnmock(defaultConfig);
 
       // Remote preset in globalExtends should be resolved
       expect(parsedConfig).toContainEntries([
@@ -207,6 +220,7 @@ describe('workers/global/config/parse/index', () => {
     });
 
     it('throws exception if global presets cannot be resolved', async () => {
+      fileConfigParser.getConfig.mockResolvedValue({});
       httpMock
         .scope('http://example.com/')
         .get('/config.json')
@@ -221,12 +235,14 @@ describe('workers/global/config/parse/index', () => {
     });
 
     it('cli dryRun replaced to full', async () => {
+      fileConfigParser.getConfig.mockResolvedValue({});
       defaultArgv = defaultArgv.concat(['--dry-run']);
       const parsed = await configParser.parseConfigs(defaultEnv, defaultArgv);
       expect(parsed).toContainEntries([['dryRun', 'full']]);
     });
 
     it('env dryRun = false replaced to null', async () => {
+      fileConfigParser.getConfig.mockResolvedValue({});
       const env: NodeJS.ProcessEnv = {
         ...defaultEnv,
         RENOVATE_DRY_RUN: 'false',
@@ -236,35 +252,31 @@ describe('workers/global/config/parse/index', () => {
     });
 
     it('cli dryRun = false replaced to null', async () => {
+      fileConfigParser.getConfig.mockResolvedValue({});
       defaultArgv = defaultArgv.concat(['--dry-run=false']);
       const parsed = await configParser.parseConfigs(defaultEnv, defaultArgv);
       expect(parsed).toContainEntries([['dryRun', null]]);
     });
 
     it('only initializes the file when the env var LOG_FILE is properly set', async () => {
-      vi.doMock('../../../../../config.js', () => ({ default: {} }));
-      const env: NodeJS.ProcessEnv = {};
-      const parsedConfig = await configParser.parseConfigs(env, defaultArgv);
+      fileConfigParser.getConfig.mockResolvedValue({});
+      const parsedConfig = await configParser.parseConfigs({}, defaultArgv);
       expect(parsedConfig).not.toContain([['logFile', 'someFile']]);
       expect(getParentDir).not.toHaveBeenCalled();
     });
 
     it('massage onboardingNoDeps when autodiscover is false', async () => {
-      vi.doMock('../../../../../config.js', () => ({
-        default: {
-          onboardingNoDeps: 'auto',
-          autodiscover: false,
-        },
-      }));
+      fileConfigParser.getConfig.mockResolvedValueOnce({
+        onboardingNoDeps: 'auto',
+        autodiscover: false,
+      });
       const env: NodeJS.ProcessEnv = {};
       const parsedConfig = await configParser.parseConfigs(env, defaultArgv);
       expect(parsedConfig).toContainEntries([['onboardingNoDeps', 'enabled']]);
     });
 
     it('apply secrets to global config', async () => {
-      vi.doMock('../../../../../config.js', () => ({
-        default: {},
-      }));
+      fileConfigParser.getConfig.mockResolvedValue({});
       const env: NodeJS.ProcessEnv = {
         ...defaultEnv,
         RENOVATE_SECRETS: '{"SECRET_TOKEN": "secret_token"}',
